+++ src/app.ts
@@ -1,208 +1,10 @@
-/// <reference path="d-d-interfaces.ts" />
-/// <reference path="project-model.ts" />
+/// <reference path="models/d-d-interfaces.ts" />
+/// <reference path="models/project-model.ts" />
+/// <reference path="state/project-state.ts" />
+/// <reference path="util/validation.ts" />
+/// <reference path="decorators/autobind-decorator.ts" />
 
 namespace App {
-    // Project Type
-
-// Project Type
-
-// Project State Mgmt class HERE
-// items will be listed here whenever add project button is clicked
-type Listener<T> = (items: Project[]) => void;
-
-class State<T> {
-    protected listeners: Listener<T>[] = [];
-
-    addListener(listenerFn: Listener<T>) {
-        this.listeners.push(listenerFn);
-    }
-}
-
-class ProjectState extends State<Project> {
-    
-    private projects: Project[] = [];
-    private static instance: ProjectState;
-
-    private constructor() {
-        super();
-    }
-// adding a static method to check if the instance exists and if it does... return it 
-    static getInstance() {
-        if (this.instance) {
-            return this.instance;
-        }
-        this.instance = new ProjectState();
-        return this.instance;
-    }
-    
-    addProject(title: string, description: string, numOfPeople: number) {
-        const newProject = new Project(
-            Math.random().toString(),
-            title,
-            description,
-            numOfPeople,
-            ProjectStatus.Active
-        );
-        this.projects.push(newProject);
-        for (const listenerFn of this.listeners) {
-            listenerFn(this.projects.slice());
-        }
-    }
-// updates the object in the array
-    moveProject(projectId: string, newStatus: ProjectStatus) {
-        const project = this.projects.find(prj => prj.id === projectId);
-        if (project) {
-            project.status = newStatus;
-            this.updateListeners();
-        }
-    }
-
-    private updateListeners() {
-        for (const listenerFn of this.listeners) {
-            listenerFn(this.projects.slice());
-        }
-    }
-}
-
-const projectState = ProjectState.getInstance();
-
-// Validation logic 
-interface Validatable {
-    value: string | number;
-    required?: boolean;
-    minLength?: number;
-    maxLength?: number;
-    min?: number;
-    max?: number;
-}
-// if the required check is flagged, we check is the value is empty etc.
-function validate(validatableInput: Validatable) {
-    let isValid = true;
-    if (validatableInput.required) {
-        isValid = isValid && validatableInput.value.toString().trim().length !== 0;
-    }
-    if (
-        validatableInput.minLength != null && 
-        typeof validatableInput.value === 'string'
-        ) {
-        isValid = 
-        isValid && validatableInput.value.length > validatableInput.minLength;
-    }
-    if (
-        validatableInput.maxLength != null &&
-        typeof validatableInput.value === 'string'
-    ) {
-        isValid = 
-        isValid && validatableInput.value.length < validatableInput.maxLength;
-    }
-    //checking the final min/max variables of the interface
-    if (validatableInput.min != null && typeof validatableInput.value === 'number') {
-        isValid = isValid && validatableInput.value > validatableInput.min;
-    }
-    if (validatableInput.max != null && typeof validatableInput.value === 'number') {
-        isValid = isValid && validatableInput.value < validatableInput.max;
-    }
-    return isValid;
-}
-
-//adding the autobind decorator so we don't need to use the bind keyword
-function autobind(
-    _: any, 
-    _2: string, 
-    descriptor: PropertyDescriptor
-    ) {
-        const originalMethod = descriptor.value;
-        const adjDescriptor: PropertyDescriptor = {
-            configurable: true,
-            get() {
-                const boundFn = originalMethod.bind(this);
-                return boundFn;
-            }
-        };
-        return adjDescriptor;
-    }
-
-// Component Base Class
-abstract class Component<T extends HTMLElement, U extends HTMLElement> {
-    templateElement: HTMLTemplateElement;
-    hostElement: T;
-    element: U;
-
-    constructor(
-        templateId: string, 
-        hostElementId: string,
-        insertAtStart: boolean, 
-        newElementId?: string
-) {
-    this.templateElement = document.getElementById(
-        templateId
-    )! as HTMLTemplateElement;
-    this.hostElement = document.getElementById(hostElementId)! as T;
-
-    const importedNode = document.importNode(
-        this.templateElement.content, 
-        true
-    );
-    this.element = importedNode.firstElementChild as U;
-    if (newElementId) {
-    this.element.id = newElementId;
-        }
-    this.attach(insertAtStart);
-    }
-
-    private attach(insertAtBeginning: boolean) {
-        this.hostElement.insertAdjacentElement(
-            insertAtBeginning ? 'afterbegin' : 'beforeend', this.element
-        );
-    }
-
-    abstract configure(): void;
-    abstract renderContent(): void;
-}
-//Project Item Class
-class ProjectItem extends Component<HTMLUListElement, HTMLLIElement>
-    implements Draggable {
-    private project: Project;
-
-    get persons() {
-        if (this.project.people === 1) {
-            return '1 person';
-        } else {
-            return `${this.project.people} persons`;
-        }
-    }
-
-    constructor(hostId: string, project: Project) {
-        super('single-project', hostId, false, project.id);
-        this.project = project;
-
-        this.configure();
-        this.renderContent();
-    }
-
-    @autobind
-    dragStartHandler(event: DragEvent) {
-        event.dataTransfer!.setData('text/plain', this.project.id);
-        event.dataTransfer!.effectAllowed = 'move';
-    }
-
-    dragEndHandler(event: DragEvent) {
-        console.log('DragEnd');
-    }
-
-    configure() {
-        this.element.addEventListener('dragstart', this.dragStartHandler);
-    }
-
-
-//setting its own content here
-    renderContent() {
-        this.element.querySelector('h2')!.textContent = this.project.title;
-        this.element.querySelector('h3')!.textContent = this.persons + ' assigned';
-        this.element.querySelector('p')!.textContent = this.project.description;
-    }
-
-}
 
 // Project List HERE
 class ProjectList extends Component<HTMLDivElement, HTMLElement> 
@@ -271,81 +73,7 @@ class ProjectList extends Component<HTMLDivElement, HTMLElement>
     }
 }
 
-//Project Input Class
-class ProjectInput extends Component<HTMLDivElement, HTMLFormElement> {
-    titleInputElement: HTMLInputElement;
-    descriptionInputElement: HTMLInputElement;
-    peopleInputElement: HTMLInputElement;
 
-    constructor() {
-        super('project-input', 'app', true, 'user-input');
-        // here populating the elements listed above
-        this.titleInputElement = this.element.querySelector('#title') as HTMLInputElement;
-        this.descriptionInputElement = this.element.querySelector('#description') as HTMLInputElement;
-        this.peopleInputElement = this.element.querySelector('#people') as HTMLInputElement;
-        this.templateElement = document.getElementById('project-input')! as HTMLTemplateElement;
-        this.configure();
-    }
-    
-    configure() {
-        this.element.addEventListener('submit', this.submitHandler);
-    }
-    
-    renderContent() {}
-
-    private gatherUserInput(): [string, string, number] | void {
-        const enteredTitle = this.titleInputElement.value;
-        const enteredDescription = this.descriptionInputElement.value;
-        const enteredPeople = this.peopleInputElement.value;
-
-        const titleValidatable: Validatable = {
-            value: enteredTitle,
-            required: true
-        };
-
-        const descriptionValidatable: Validatable = {
-            value: enteredDescription,
-            required: true,
-            minLength: 5
-        };
-
-        const peopleValidatable: Validatable = {
-            value: +enteredPeople,
-            required: true,
-            min: 1,
-            max: 5
-        };
-// if all these conditions are satisfied, then continue
-// revised to have a validation function that is resusable
-        if (
-            !validate(titleValidatable) ||
-            !validate(descriptionValidatable) ||
-            !validate(peopleValidatable)
-        ) {
-            alert('Invalid input, please try again!');
-            return;
-        } else {
-            return [enteredTitle, enteredDescription, +enteredPeople];
-        }
-    }
-// empty all inputs after submission
-    private clearInputs() {
-        this.titleInputElement.value = '';
-        this.descriptionInputElement.value = '';
-        this.peopleInputElement.value = '';
-    }
-
-    @autobind
-    private submitHandler(event: Event) {
-        event.preventDefault();
-        const userInput = this.gatherUserInput();
-        if (Array.isArray(userInput)) {
-            const [title, desc, people] = userInput;
-            projectState.addProject(title, desc, people);
-            this.clearInputs();
-        }
-    }
-}
 // rendering the form through this line
 new ProjectInput();
 new ProjectList('active');
