+++ src/models/d-d-interfaces.ts
